app/requirements.txt
django==3.2.2
django-jsonfield==1.4.1
gunicorn==20.0.4
djangorestframework==3.12.4
python-decouple==3.4
psycopg2-binary==2.9.1